requirements.txt
# Azure SDK
azure-search-documents==11.4.0
azure-cosmos==4.5.1
azure-storage-blob==12.17.0
azure-identity==1.15.0
openai==1.3.5

# Data processing
sentence-transformers==2.2.2
transformers==4.35.2
numpy==1.24.3
pandas==2.0.3

# Configuration and utilities
pydantic==2.5.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and monitoring
structlog==23.2.0